strip out models into separate files
