+++ controllers/topics.controllers.js
@@ -1,4 +1,4 @@
-const { selectAllTopics } = require("../models/topics.models.js");
+const { selectAllTopics } = require("../models/");
 
 exports.getTopics = (req, res, next) => {
     selectAllTopics()
