+++ models/index.js
@@ -1,0 +1,6 @@
+module.exports = {
+    ...require("./articles.models"),
+    ...require("./comments.models"),
+    ...require("./endpoints.models"),
+    ...require("./topics.models")
+}
